# Remove transformer from DI