Avoid DI for `AddSchemaDescriptionsTransformer` as it seems to have a non-negligible performance penalty.