@martincostello 2024-08-07
Before: https://github.com/martincostello/aspnetcore-openapi/pull/22#issuecomment-2273110771

After:

```

BenchmarkDotNet v0.14.0, Windows 11 (10.0.22621.3880/22H2/2022Update/SunValley2)
12th Gen Intel Core i7-1270P, 1 CPU, 16 logical and 12 physical cores
.NET SDK 9.0.100-preview.7.24406.8
  [Host]   : .NET 9.0.0 (9.0.24.40507), X64 RyuJIT AVX2
  ShortRun : .NET 9.0.0 (9.0.24.40507), X64 RyuJIT AVX2

Job=ShortRun  IterationCount=3  LaunchCount=1  
WarmupCount=3  

```
| Method      | Mean        | Error        | StdDev      | Median      | Gen0     | Gen1     | Gen2     | Allocated  |
|------------ |------------:|-------------:|------------:|------------:|---------:|---------:|---------:|-----------:|
| AspNetCore  | 18,696.7 μs | 140,804.8 μs | 7,717.99 μs | 14,427.2 μs | 609.3750 | 406.2500 | 156.2500 | 5290.23 KB |
| NSwag       |    373.4 μs |   1,049.7 μs |    57.54 μs |    367.1 μs |   2.9297 |        - |        - |   26.56 KB |
| Swashbuckle | 11,594.9 μs |   1,700.9 μs |    93.23 μs | 11,565.3 μs | 156.2500 |  31.2500 |        - |  1543.7 KB |